+++ common/src/main/java/com/viaversion/viabackwards/protocol/v1_11to1_10/Protocol1_11To1_10.java
@@ -23,7 +23,7 @@ import com.viaversion.viabackwards.api.data.BackwardsMappingData;
 import com.viaversion.viabackwards.api.rewriters.SoundRewriter;
 import com.viaversion.viabackwards.protocol.v1_11to1_10.rewriter.BlockItemPacketRewriter1_11;
 import com.viaversion.viabackwards.protocol.v1_11to1_10.rewriter.EntityPacketRewriter1_11;
-import com.viaversion.viabackwards.protocol.v1_11to1_10.rewriter.PlayerPacketRewriterRewriter1_11;
+import com.viaversion.viabackwards.protocol.v1_11to1_10.rewriter.PlayerPacketRewriter1_11;
 import com.viaversion.viabackwards.protocol.v1_11to1_10.storage.WindowTracker;
 import com.viaversion.viaversion.api.connection.UserConnection;
 import com.viaversion.viaversion.api.minecraft.ClientWorld;
@@ -35,8 +35,8 @@ import com.viaversion.viaversion.protocols.v1_9_1to1_9_3.packet.ServerboundPacke
 public class Protocol1_11To1_10 extends BackwardsProtocol<ClientboundPackets1_9_3, ClientboundPackets1_9_3, ServerboundPackets1_9_3, ServerboundPackets1_9_3> {
 
     public static final BackwardsMappingData MAPPINGS = new BackwardsMappingData("1.11", "1.10");
-    private final EntityPacketRewriter1_11 entityPackets = new EntityPacketRewriter1_11(this);
-    private final BlockItemPacketRewriter1_11 blockItemPackets = new BlockItemPacketRewriter1_11(this);
+    private final EntityPacketRewriter1_11 entityRewriter = new EntityPacketRewriter1_11(this);
+    private final BlockItemPacketRewriter1_11 itemRewriter = new BlockItemPacketRewriter1_11(this);
 
     public Protocol1_11To1_10() {
         super(ClientboundPackets1_9_3.class, ClientboundPackets1_9_3.class, ServerboundPackets1_9_3.class, ServerboundPackets1_9_3.class);
@@ -44,9 +44,9 @@ public class Protocol1_11To1_10 extends BackwardsProtocol<ClientboundPackets1_9_
 
     @Override
     protected void registerPackets() {
-        blockItemPackets.register();
-        entityPackets.register();
-        PlayerPacketRewriterRewriter1_11.register(this);
+        entityRewriter.register();
+        itemRewriter.register();
+        PlayerPacketRewriter1_11.register(this);
 
         SoundRewriter<ClientboundPackets1_9_3> soundRewriter = new SoundRewriter<>(this);
         soundRewriter.registerNamedSound(ClientboundPackets1_9_3.CUSTOM_SOUND);
@@ -73,12 +73,12 @@ public class Protocol1_11To1_10 extends BackwardsProtocol<ClientboundPackets1_9_
 
     @Override
     public EntityPacketRewriter1_11 getEntityRewriter() {
-        return entityPackets;
+        return entityRewriter;
     }
 
     @Override
     public BlockItemPacketRewriter1_11 getItemRewriter() {
-        return blockItemPackets;
+        return itemRewriter;
     }
 
     @Override
