--- common/src/main/java/com/viaversion/viabackwards/protocol/v1_15to1_14_4/data/EntityTypeMapping.java
@@ -1,29 +1,0 @@
-/*
- * This file is part of ViaBackwards - https://github.com/ViaVersion/ViaBackwards
- * Copyright (C) 2016-2024 ViaVersion and contributors
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-package com.viaversion.viabackwards.protocol.v1_15to1_14_4.data;
-
-import com.viaversion.viaversion.api.minecraft.entities.EntityTypes1_14;
-
-public class EntityTypeMapping {
-
-    // There's only the bee, so not much to do here
-    public static int getOldEntityId(int entityId) {
-        if (entityId == 4) return EntityTypes1_14.PUFFERFISH.getId(); // Flying pufferfish!
-        return entityId >= 5 ? entityId - 1 : entityId;
-    }
-}
