+++ common/src/main/java/com/viaversion/viabackwards/protocol/v1_16to1_15_2/storage/WorldNameTracker.java
@@ -15,31 +15,18 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.viaversion.viabackwards.protocol.v1_18to1_17_1.data;
+package com.viaversion.viabackwards.protocol.v1_16to1_15_2.storage;
 
-import com.viaversion.viaversion.protocols.v1_17_1to1_18.data.BlockEntityIds1_18;
-import java.util.Arrays;
+import com.viaversion.viaversion.api.connection.StorableObject;
 
-public final class BlockEntityIds {
+public class WorldNameTracker implements StorableObject {
+    private String worldName;
 
-    private static final int[] IDS;
-
-    static {
-        final int[] ids = BlockEntityIds1_18.getIds();
-        IDS = new int[Arrays.stream(ids).max().getAsInt() + 1];
-        Arrays.fill(IDS, -1);
-        for (int i = 0; i < ids.length; i++) {
-            final int id = ids[i];
-            if (id != -1) {
-                IDS[id] = i;
-            }
-        }
+    public String getWorldName() {
+        return worldName;
     }
 
-    public static int mappedId(final int id) {
-        if (id < 0 || id > IDS.length) {
-            return -1;
-        }
-        return IDS[id];
+    public void setWorldName(String worldName) {
+        this.worldName = worldName;
     }
 }
