+++ common/src/main/java/com/viaversion/viabackwards/protocol/v1_9_3to1_9_1/data/BlockEntity1_9_1.java
@@ -15,12 +15,11 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.viaversion.viabackwards.protocol.v1_9_3to1_9_1.chunks;
+package com.viaversion.viabackwards.protocol.v1_9_3to1_9_1.data;
 import com.viaversion.viabackwards.protocol.v1_9_3to1_9_1.Protocol1_9_3To1_9_1;
 import com.viaversion.viaversion.api.connection.UserConnection;
 import com.viaversion.viaversion.api.minecraft.Position;
 import com.viaversion.viaversion.api.protocol.packet.PacketWrapper;
-import com.viaversion.viaversion.api.type.Type;
 import com.viaversion.viaversion.api.type.Types;
 import com.viaversion.viaversion.protocols.v1_9_1to1_9_3.packet.ClientboundPackets1_9_3;
 import com.viaversion.nbt.tag.CompoundTag;
@@ -29,7 +28,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-public class BlockEntity {
+public class BlockEntity1_9_1 {
     private static final Map<String, Integer> TYPES = new HashMap<>();
 
     static {
